Recipe Selection Menu Recipe Desc and Photo Layout
There are four canvases in the current layout. Each one has their
specfic names. The spacing and such can be played around with after. For
the time being, I made the extra two recipes, Spring Rolls, and Ribs. We
can change this later. So far, it's just Mango Pudding Cake, Pow
Chicken, Spring Rolls and Ribs.

Oh! Added more things to the drop down menu and the title for the
recommended recipes.